Fix editdistance bug

+++ src/editdistance.h
@@ -1,7 +1,9 @@
 #ifndef EDITDISTANCE_H_
 #define EDITDISTANCE_H_
 
+#include <stdlib.h>
+
 /* calculate the edit distance between two strings */
-int editdistance (char* a, int m, char* b, int n);
+int editdistance (char* a, size_t m, char* b, size_t n);
 
 #endif
